TypeScript works with 1.0.1 now
Now supports the newest version of TypeScript by supplying
ScriptSnapshot, getSyntacticDiagnostics, and other new compilation
requirement methods introduced to the TypeScript library. Added the
variable scripts to keep track of the number of scripts being pulled and
fire the compiler after the number of loaded scripts reaches the total
number of scripts. Also now allows support for older IE as XHR request
was failing. It wasn't previously waiting until the scripts were
completely loaded sometimes before pushing them into the src and
requesting compile. Pulled the script tag scraper out of the compiler
and ran it in an anonymous function to initiate the loading of the
content before firing the compiler.